move slurm scripts into src and add use-case specific script
