+++ src/itwinai/slurm/slurm_constants.py
